Update API endpoints in environment files
The API endpoints have been updated in the .env.development and .env files for both Curator Dashboard and Submission Form applications. The changes were needed to point the applications towards our dev.curation and caltech-curation environments instead of the old mangolassi and tazendra servers for the API read and autocomplete functions. This move is part of the transition to our new curation infrastructure.