Optimize text preprocessing and enhance SentenceClassification component
This commit introduces a more elaborate preprocessing mechanism in the curator_dashboard.py, where extra considerations have been made to standardize the paper text, such as replacing phrases and removing certain sections for a cleaner text analysis.

In the SentenceClassification.js frontend component, useState hooks are added to track the classifier type and the result type, which can be toggled to filter presentation of the sentences. The classification results for "Gene Expression" and "Kinase Activity" are now presented in separate tabs for better user experience. The changes enhance usability and improve readability of the displayed data, hence increasing efficiency in the curation process.