Add SentenceClassificationAPI for sentence classification tasks
Added a new API to handle sentence classification tasks (sentence_classification_api.py) which now handles the loading of sentence classifiers and sent2vec model, and processes POST requests for classifying sentences. Refactored curator_dashboard.py to now make a POST request to this new API for sentence classifications. This allows for better separation of concerns, as sentence classifications tasks has been abstracted out of the curator dashboard. This commit also includes the Dockerfiles for both the new API and changes to old Dockerfile for the environment variable pointing to this new API.